+++ app/(interactive)/4-3/page.tsx
@@ -1,64 +1,30 @@
 "use client";
-import { getStroke } from "perfect-freehand";
-import { useContext } from "react";
-import AppContext from "../../../components/AppContext";
 import NextButton from "@/components/ui/nextButton";
+import { motion } from "framer-motion";
 import Link from "next/link";
-type Props = {};
 
-const Page = (props: Props) => {
-  const context = useContext(AppContext);
-  const strokes = context.session;
-
-  function getSvgPathFromStroke(stroke: any) {
-    if (!stroke.length) return "";
-    // reduce the size of svg by divisor
-    const divisor = 1.41;
-    const d = stroke.reduce(
-      (acc: any[], [x0, y0]: any, i: number, arr: string | any[]) => {
-        const [x1, y1] = arr[(i + 1) % arr.length];
-        acc.push(
-          x0 / divisor,
-          y0 / divisor,
-          (x0 + x1) / 2 / divisor,
-          (y0 + y1) / 2 / divisor,
-        );
-        return acc;
-      },
-      ["M", stroke[0][0] / divisor, stroke[0][1] / divisor, "Q"],
-    );
-    d.push("Z");
-    return d.join(" ");
-  }
-
-  const renderedStrokes = strokes.map((stroke: any, index: any) => {
-    const pathData = getSvgPathFromStroke(
-      getStroke(stroke, {
-        size: 8,
-        thinning: 0.5,
-        smoothing: 0.5,
-        streamline: 0.5,
-      }),
-    );
-    return <path key={index} d={pathData} />;
-  });
+export default function Page() {
+  const text_4_3 = '"นี่มันไดอารี่ของเรานี่นา"';
 
   return (
-    <div className="flex h-full w-full flex-col items-center">
-      <div className="fixed top-[140px] flex items-center space-x-[16px] text-center">
-        <svg
-          id="svg"
-          className="relative h-[112px] w-[235px] touch-none rounded-[10px] bg-[#F8F8F7] opacity-80"
-        >
-          {renderedStrokes}
-        </svg>
-        <p className="text[#1E1B20] font-cursive text-[40px]">'s</p>
-      </div>
-      <Link href="/4-4">
-      <NextButton/>
+    <div className="text-white text-center mx-auto min-w-[430px] min-h-screen bg-4-3 bg-cover bg-no-repeat">
+      <motion.div
+        initial={{ opacity: 0, z: -20 }}
+        animate={{ opacity: 1, z: 0, transition: { duration: 1, delay: 1 } }}
+        className="mt-[218px] mb-[403px]"
+      >
+        <p>{text_4_3}</p>
+      </motion.div>
+      
+
+      <Link href="4-4">
+      <motion.div
+        initial={{ opacity: 0, z: -20 }}
+        animate={{ opacity: 1, z: 0, transition: { duration: 1, delay: 2.5 } }}
+      >
+        <NextButton />
+      </motion.div>
       </Link>
     </div>
   );
-};
-
-export default Page;
+}
