+++ app/(interactive)/4-6/page.tsx
@@ -1,10 +1,147 @@
-import React from 'react'
+"use client";
+import NextButton from "@/components/ui/nextButton";
+import Link from "next/link";
+import { getStroke } from "perfect-freehand";
+import React, { useContext, useEffect, useRef } from "react";
+import AppContext from "../../../components/AppContext";
 
-export default function page() {
+type Props = {};
+
+const Page = (props: Props) => {
+  function getSvgPathFromStroke(stroke: any) {
+    if (!stroke.length) return "";
+    const d = stroke.reduce(
+      (acc: any[], [x0, y0]: any, i: number, arr: string | any[]) => {
+        const [x1, y1] = arr[(i + 1) % arr.length];
+        acc.push(x0, y0, (x0 + x1) / 2, (y0 + y1) / 2);
+        return acc;
+      },
+      ["M", ...stroke[0], "Q"],
+    );
+    d.push("Z");
+    return d.join(" ");
+  }
+
+  const [hasWrite, setWrite] = React.useState<boolean>(false);
+  const [strokes, setStrokes] = React.useState<number[][][]>([]);
+  const currentStroke = useRef<number[][]>([]);
+  const context = useContext(AppContext);
+  const svgRef = useRef();
+
+  useEffect(() => {
+    return () => {
+      context.setSession(strokes);
+    };
+  }, [strokes]);
+
+  function handlePointerDown(e: React.PointerEvent<SVGSVGElement>) {
+    setWrite(true);
+    e.currentTarget.setPointerCapture(e.pointerId);
+    const pos = svgRef.current.getBoundingClientRect();
+    currentStroke.current = [
+      [e.pageX - pos.left, e.pageY - pos.top, e.pressure],
+    ];
+  }
+
+  function handlePointerMove(e: React.PointerEvent<SVGSVGElement>) {
+    if (e.buttons !== 1) return;
+    const pos = svgRef.current.getBoundingClientRect();
+    currentStroke.current.push([
+      e.pageX - pos.left,
+      e.pageY - pos.top,
+      e.pressure,
+    ]);
+    setStrokes([...strokes]);
+  }
+
+  function handlePointerUp() {
+    // new function to handle new independent stroke
+    setStrokes([...strokes, currentStroke.current]);
+    currentStroke.current = [];
+  }
+
+  function clearStrokes() {
+    setWrite(false);
+    setStrokes([]);
+  }
+
+  const renderedStrokes = strokes.map((stroke, index) => {
+    const pathData = getSvgPathFromStroke(
+      getStroke(stroke, {
+        size: 8,
+        thinning: 0.5,
+        smoothing: 0.5,
+        streamline: 0.5,
+      }),
+    );
+    return <path key={index} d={pathData} />;
+  });
+
   return (
-    <div className="ml-[87px] mt-[232px]">
-      <p>เรื่องที่อยากทำ</p>
-      <p className="mt-3 text-center">เรื่องที่ยังทำไม่ได้</p>
-    </div>
-  )
-}
+    <>
+      <div className=" mx-auto min-w-[430px] min-h-screen bg-4-5 bg-cover bg-no-repeat">
+        <div className="mt-[231px] mb-[150px] flex flex-col items-center text-center">
+          <p className="text-white mb-4">เจ้าของสมุดเล่มนี้คือ...</p>
+          <div className="mt-20 relative">
+            <svg
+              id="svg"
+              ref={svgRef}
+              onPointerDown={handlePointerDown}
+              onPointerMove={handlePointerMove}
+              onPointerUp={handlePointerUp}
+              className="relative h-[160px] w-[342px] touch-none rounded-[10px] bg-[#E8E5DD] opacity-100"
+            >
+              {renderedStrokes}
+              {currentStroke.current.length > 0 && (
+                <path
+                  d={getSvgPathFromStroke(
+                    getStroke(currentStroke.current, {
+                      size: 8,
+                      thinning: 0.5,
+                      smoothing: 0.5,
+                      streamline: 0.5,
+                    }),
+                  )}
+                />
+              )}
+            </svg>
+            {!hasWrite && (
+              <p className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 transform text-center text-[#B4A49A]">
+                ลากนิ้วเพื่อเขียนชื่อ
+              </p>
+            )}
+
+            <div
+              onClick={clearStrokes}
+              className="absolute left-[285px] top-[11px]"
+            >
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                width="35"
+                height="34"
+                viewBox="0 0 24 24"
+                fill="none"
+                stroke="#1E1B20"
+                strokeWidth="2"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                className="lucide lucide-eraser"
+              >
+                <path d="m7 21-4.3-4.3c-1-1-1-2.5 0-3.4l9.6-9.6c1-1 2.5-1 3.4 0l5.6 5.6c1 1 1 2.5 0 3.4L13 21" />
+                <path d="M22 21H7" />
+                <path d="m5 11 9 9" />
+              </svg>
+            </div>
+          </div>
+          <br />
+        </div>
+
+        <Link href="/4-7">
+          <NextButton></NextButton>
+        </Link>
+      </div>
+    </>
+  );
+};
+
+export default Page;
