+++ app/(interactive)/4-8/page.tsx
@@ -1,0 +1,13 @@
+"use client";
+import NextButton from "@/components/ui/nextButton";
+import { motion } from "framer-motion";
+import Link from "next/link";
+
+export default function Page() {
+  return (
+    <Link href="/4-6">
+    <div className="mx-auto min-w-[430px] min-h-screen bg-4-3 bg-cover bg-no-repeat">
+    </div>
+    </Link>
+  );
+}
