+++ package.json
@@ -13,6 +13,7 @@
     "@radix-ui/react-slot": "^1.0.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.0",
+    "framer-motion": "^10.18.0",
     "lucide-react": "^0.292.0",
     "next": "14.0.1",
     "perfect-freehand": "^1.2.0",
